refactor: create summer school template
now adding new summer school page create via
adding new object to python dict
