+++ src/templates/summer_school.html
@@ -1,9 +1,9 @@
 {% extends "base_light.html" %}
-{% block title %}Летняя школа на мат-мехе СПбГУ (2022){% endblock %}
-{% block description %}Летняя проектная онлайн-школа программирования для обучающихся СПбГУ (2022, онлайн). Школа проводилась с 4 по 30 июля 2022 года.{% endblock %}
+{% block title %}Летняя школа на мат-мехе СПбГУ (2024){% endblock %}
+{% block description %}Летняя проектная онлайн-школа программирования для обучающихся СПбГУ (2024, онлайн). Школа проводилась с 8 июля по 3 августа 2024 года.{% endblock %}
 
 {% block headers %}
-<link rel="canonical" href="https://se.math.spbu.ru/summer_school_2022.html" />
+<link rel="canonical" href="https://se.math.spbu.ru/{{school.href}}.html" />
 {% endblock %}
 
 {% block content %}
@@ -16,9 +16,9 @@
                     <div class="col-lg-9">
                         <a href="{{ url_for('summer_school_list') }}">Летние школы других годов</a>
                         <!-- Heading -->
-                        <h1 class="h2 mb-2">Летняя проектная онлайн-школа программирования для обучающихся СПбГУ (<strong>2022</strong>, онлайн).</h1>
+                        <h1 class="h2 mb-2">{{ school.full_name|safe }}</h1>
                         <!-- Text -->
-                        <p class="lead">Школа проводилась с <strong>4 по 30 июля</strong>. 4 июля в 12:00 прошло открытие школы и презентация проектов. 30 июля были проведены презентации результатов проектов, подведение итогов и закрытие школы.</p>
+                        <p class="lead">{{ school.description|safe }}</p>
                     </div>
                 </div>
             </div>
@@ -75,6 +75,7 @@
                 {% endfor %}
             </div>
         </div>
-    </section>
+</section>
+
 
 {% endblock %}
